Demo checksum updates
Verify checksums, move checksum logic into separate file, fix bug where
checksums weren't added to automatically continued recordings
