+++ src/Checksum.hpp
@@ -1,0 +1,4 @@
+#pragma once
+
+bool AddDemoChecksum(const char* filename);
+int VerifyDemoChecksum(const char* filename);
