+++ src/SourceAutoRecord.vcxproj
@@ -121,6 +121,7 @@
     <ClCompile Include="..\lib\minhook\hook.c" />
     <ClCompile Include="..\lib\minhook\trampoline.c" />
     <ClCompile Include="Cheats.cpp" />
+    <ClCompile Include="Checksum.cpp" />
     <ClCompile Include="Command.cpp" />
     <ClCompile Include="Features\ClassDumper.cpp" />
     <ClCompile Include="Features\Config.cpp" />
@@ -313,6 +314,7 @@
     <ClInclude Include="..\lib\SFML\include\SFML\System\Vector2.hpp" />
     <ClInclude Include="..\lib\SFML\include\SFML\System\Vector3.hpp" />
     <ClInclude Include="Cheats.hpp" />
+    <ClInclude Include="Checksum.hpp" />
     <ClInclude Include="Command.hpp" />
     <ClInclude Include="Features.hpp" />
     <ClInclude Include="Features\ClassDumper.hpp" />
